requirements.txt
discord.py
flask
PyNaC1
dnspython
